Refactored folder structures
part 1 of refactoring kay nay problema
