+++ client/src/App.js
@@ -16,20 +16,20 @@ import SignUp from './pages/public/SignUp';
 
 /* Student Pages */
 import StudentLayout from './components/layout/StudentLayout';
-import Documents from './pages/student/documents/Documents';
+import Documents from './pages/student/Documents';
 import Enrollment from './pages/student/Enrollment';
-import Grades from './pages/student/grades/Grades';
-import Home from './pages/student/home/Home';
+import Grades from './pages/student/Grades';
+import Home from './pages/student/Home';
 import StudentSchedule from './pages/student/StudentSchedule';
 
 /* Teacher Pages */
 import TeacherLayout from './components/layout/TeacherLayout';
-import TeacherGrades from './pages/teacher/grades/Grades';
-import TeacherHome from './pages/teacher/home/Home';
+import TeacherGrades from './pages/teacher/Grades';
+import TeacherHome from './pages/teacher/Home';
 
 /* Teacher Pages */
 import AdminLayout from './components/layout/AdminLayout';
-import AdminHome from './pages/admin/home/Home';
+import AdminHome from './pages/admin/Home';
