+++ client/src/pages/public/PaymentForm.js
@@ -29,7 +29,7 @@ function PaymentForm() {
         <section className='flex justify-start bg-white-yellow-tone bg-center bg-cover bg-repeat bg-blend-multiply' style={{ backgroundImage: `url(${Bg_image})`, minHeight: '100vh'}}>
             <SignUpNav/>
             <div className='flex flex-col justify-center items-center w-full'>
-                <div className="relative bg-white-yellow-tone w-11/12 px-8 py-4 mt-24 mb-2 mb-auto rounded-xl border-2 border-dark-red shadow-2xl md:min-h-[580px] h-auto flex flex-col">
+                <div className="relative bg-white-yellow-tone w-11/12 px-8 py-4 mt-24 mb-auto rounded-xl border-2 border-dark-red shadow-2xl md:min-h-[580px] h-auto flex flex-col">
                     <BackButton onClick={navigateToLogin} />
                         <form className="px-8 py-4 flex flex-col">
                             <h1 className='text-center text-3xl font-bold -mb-2'>Maya Payment Form</h1>
